Refactored code to match Clippy warnings

+++ src/cmd/utilcmd.rs
@@ -32,41 +32,38 @@ command!(purge(_ctx, msg, _args, amount: u64) {
 });
 
 fn purge_messages(_ctx: &mut Context, msg: &Message, amount: u64, user: Option<UserId>) {
-     match msg.channel_id.messages(|g| g.limit(100)) {
-        Ok(x) => {
-            let filtered_result: Vec<MessageId> = x.into_iter().filter(|p| {
-                if let Some(uid) = user {
-                    if p.author.id != uid {
-                        return false;
-                    }
-                }
-                let ptime = p.timestamp.clone();
-                let ctime = Local::now();
-                let delta = ctime.signed_duration_since(ptime);
-                if delta.num_days() > 11 {
+    if let Ok(x) = msg.channel_id.messages(|g| g.limit(100)) {
+        let filtered_result: Vec<MessageId> = x.into_iter().filter(|p| {
+            if let Some(uid) = user {
+                if p.author.id != uid {
                     return false;
                 }
-                true
-
-            }).map(|i| i.id).take((amount + 1) as usize).collect();
-            if filtered_result.len() > 1 {
-                match msg.channel_id.delete_messages(&filtered_result) {
-                    Ok(_) => {
-                        debug!("Purge of {} OK !", msg.channel_id);
-                        let _ = msg.delete();
-                    },
-                    Err(e) => warn!("Purge of {} ERR !\nARG (LENGTH: {}):  {:?}\nERR: {:?}", msg.channel_id, filtered_result.len(), filtered_result, e)
-                }
-            } else if filtered_result.len() == 1 {
-                match msg.channel_id.delete_message(filtered_result[0]) {
-                    Ok(_) => {
-                        debug!("Purge of {} OK !", msg.channel_id);
-                        let _ = msg.delete();
-                    },
-                    Err(e) => warn!("Purge of {} ERR !\nARG (LENGTH: {}):  {:?}\nERR: {:?}", msg.channel_id, filtered_result.len(), filtered_result, e)
-                }
             }
-        },
-        Err(_) => ()
+            let ptime = p.timestamp;
+            let ctime = Local::now();
+            let delta = ctime.signed_duration_since(ptime);
+            if delta.num_days() > 11 {
+                return false;
+            }
+            true
+
+        }).map(|i| i.id).take((amount + 1) as usize).collect();
+        if filtered_result.len() > 1 {
+            match msg.channel_id.delete_messages(&filtered_result) {
+                Ok(_) => {
+                    debug!("Purge of {} OK !", msg.channel_id);
+                    let _ = msg.delete();
+                },
+                Err(e) => warn!("Purge of {} ERR !\nARG (LENGTH: {}):  {:?}\nERR: {:?}", msg.channel_id, filtered_result.len(), filtered_result, e)
+            }
+        } else if filtered_result.len() == 1 {
+            match msg.channel_id.delete_message(filtered_result[0]) {
+                Ok(_) => {
+                    debug!("Purge of {} OK !", msg.channel_id);
+                    let _ = msg.delete();
+                },
+                Err(e) => warn!("Purge of {} ERR !\nARG (LENGTH: {}):  {:?}\nERR: {:?}", msg.channel_id, filtered_result.len(), filtered_result, e)
+            }
+        }
     }
 }
